# feat(refactor): refactor toESDocument() to be more efficient