refactor code for performance & readability

this PR removes the slow JSON serialize/deserialize code. that code converts each object to a string and then back in to an object in order to take advantage of a feature in `JSON.stringify` which removes empty properties.

it turns out this is potentially slow, so I refactored the code to do the same thing using a different method.

benchmarks:
```bash
#### calling constructor, all setter methods and toESDocument()

# before (after the previous PR was merged)
imported 10000 docs in 1124 ms
average speed: 0.1124 ms

# after
imported 10000 docs in 290 ms
average speed: 0.0290 ms
```

works out to be ~75% faster, multiplied out to 500 million records, this change could save 10+ hrs